Move webappoauth to lib
